+++ handlers/handle_cache.ts
@@ -1,27 +1,15 @@
-import { green } from "https://deno.land/std/fmt/colors.ts";
-import { v4 } from "https://deno.land/std/uuid/mod.ts";
+import { green, red } from "https://deno.land/std/fmt/colors.ts";
+import { needProxy, Proxy } from "../proxy/proxy.ts";
 import { STD } from "../utils/info.ts";
-
 import db from "../utils/db.ts";
 
-const id = v4.generate();
-
-// * generate universally unique identifier
-function generateCacheId(): string {
-  // * standardize the id to be a valid name
-  const hash = id.split("-");
-  hash.unshift("Trex_Cache_");
-
-  return hash.join("");
-}
-
 async function cached(typePkg: string, packageUrl: string) {
-  const ID = generateCacheId();
+  const ID = "Trex_Cache_Map";
 
   let process: Deno.Process;
 
   console.log(green("cache package... \n"));
-  // *
+
   if (STD.includes(typePkg) && db.includes(typePkg)) {
     process = Deno.run({
       cmd: [
@@ -31,7 +19,7 @@ async function cached(typePkg: string, packageUrl: string) {
         "-n",
         ID,
         "--unstable",
-        packageUrl + "mod.ts",
+        needProxy(typePkg) ? Proxy(typePkg) : packageUrl + "mod.ts",
       ],
     });
   }
@@ -45,33 +33,26 @@ async function cached(typePkg: string, packageUrl: string) {
         "-n",
         ID,
         "--unstable",
-        packageUrl + "mod.ts",
+        needProxy(typePkg) ? Proxy(typePkg) : packageUrl + "mod.ts",
       ],
     });
 
     await process.status();
-    console.log(green("\n Done \n"));
+    console.log(green("\n Done. \n"));
   }
   // * install third party package
   else if (db.includes(typePkg)) {
     process = Deno.run({
-      cmd: [
-        "deno",
-        "install",
-        "-f",
-        "-n",
-        ID,
-        "--unstable",
-        packageUrl,
-      ],
+      cmd: ["deno", "install", "-f", "-n", ID, "--unstable", packageUrl],
     });
 
     await process.status();
-    console.log(green("\n Done \n"));
+    console.log(green("\n Done. \n"));
   }
-  // * throw error if package is not found
+  // * log error if package is not found
   else if (!STD.includes(typePkg) && !db.includes(typePkg)) {
-    throw new TypeError("package not found");
+    console.error(red("package not found"));
+    return;
   }
 }